+++ common/src/main/kotlin/PackageInfo.kt
@@ -1,1 +1,0 @@
-package dev.inmo.plagubot.plugins.common