+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 0.6.2
+
+* Fixes in thread-related databases
+
 ## 0.6.1
 
 * `Versions`: